feat(backend,lower,kernel,mirage,tvm): backend interface, param() migration, KP cleanup + migration
Evolve the Backend vtable: buffer_to_host returns ?Event (null for
 synchronous/host-resident transfers), execute_into takes []Buffer
 instead of []?RawBuffer, add serialize_executable/load_serialized,
 add transfer() convenience for HostBuffer and Tree(HostBuffer)
 upload. PJRT and IREE backends implement the new signatures.

Migrate all param_* accessor calls in lowering and kernelization to
 the generic param(comptime tag, params). Add f16/i8/u8 arms to every
 DType switch in lowering (dtype_to_mlir_type, scalar_zero_bytes, etc),
 PJRT (BufferType, ffi_dtype_to_kernel_dtype, dtype_to_buffer_type),
 and IREE (dtype_to_element_type).

Remove MlirKernelDescriptor and compile_mlir_fn from the kernel
 provider interface, MLIR-level kernelization types move into
 mirage/mlir.zig as provider-internal concerns.

+++ src/tvm/module.zig
@@ -202,9 +202,9 @@ pub fn load_cached(
 pub fn matmul_cache_key(
     allocator: std.mem.Allocator,
     target_kind: TargetKind,
-    m: usize,
-    n: usize,
-    k: usize,
+    m: i64,
+    n: i64,
+    k: i64,
 ) ![]u8 {
     const target_suffix = switch (target_kind) {
         .cpu => "cpu",
@@ -327,7 +327,7 @@ fn parse_run_secs(line: []const u8) ?f64 {
     while (i > 4) {
         i -= 1;
         if (line[i] == '{' and line[i - 1] == ',' and line[i - 2] == ']') {
-            // Found ],{ — now find the opening [ of the run_secs array
+            // Found ],{ now find the opening [ of the run_secs array
             var j = i - 3;
             while (j > 0 and line[j] != '[') j -= 1;
             if (j > 0 and line[j] == '[') {
